requirements.txt
torch 
torchvision 
torchaudio
transformers
sentencepiece
huggingface-hub
hf_xet
ipywidgets
sentence-transformers
ollama
pgvector
psycopg[binary]
datasets
bitsandbytes
fastapi
uvicorn
jinja2
bs4
lxml
python-dotenv
